+++ .eslintrc.js
@@ -30,6 +30,9 @@ module.exports = {
   overrides: [
     {
       files: ['*.js'],
+      rules: {
+        '@typescript-eslint/no-var-requires': 'off', // Disable the rule for JavaScript files
+      },
     },
   ],
 };
