+++ lib/stripe.ts
@@ -1,8 +1,8 @@
-import { createStripeTeam, getByTeamId } from 'models/stripeTeam';
 import Stripe from 'stripe';
 import env from '@/lib/env';
+import { updateTeam } from 'models/team';
 
-export const stripe = new Stripe(env.stripe.stripeSecretKey ?? '', {
+export const stripe = new Stripe(env.stripe.secretKey ?? '', {
   // https://github.com/stripe/stripe-node#configuration
   apiVersion: '2022-11-15',
   // Register this as an official Stripe plugin.
@@ -13,19 +13,9 @@ export const stripe = new Stripe(env.stripe.stripeSecretKey ?? '', {
   },
 });
 
-export const getURL = () => {
-  let url = env.appUrl || 'http://localhost:4002/';
-  // Make sure to include `https://` when not localhost.
-  url = url.includes('http') ? url : `https://${url}`;
-  // Make sure to including trailing `/`.
-  url = url.charAt(url.length - 1) === '/' ? url : `${url}/`;
-  return url;
-};
-
 export async function getStripeCustomerId(teamMember, session?: any) {
   let customerId = '';
-  const stripeTeam = await getByTeamId(teamMember.teamId);
-  if (!stripeTeam) {
+  if (!teamMember.team.stripeCustomerId) {
     const customerData: {
       metadata: { teamId: string };
       email?: string;
@@ -39,10 +29,12 @@ export async function getStripeCustomerId(teamMember, session?: any) {
       ...customerData,
       name: session?.user?.name as string,
     });
-    await createStripeTeam(teamMember.teamId, customer.id);
+    await updateTeam(teamMember.team.slug, {
+      stripeCustomerId: customer.id,
+    });
     customerId = customer.id;
   } else {
-    customerId = stripeTeam.customerId;
+    customerId = teamMember.team.stripeCustomerId;
   }
   return customerId;
 }
